+++ modules/tinymce/CHANGELOG.md
@@ -14,15 +14,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The RGB fields in the color picker dialog were not staying in sync with the color palette and hue slider #TINY-6952
 - The color preview box in the color picker dialog was not correctly displaying the saturation and value of the chosen color #TINY-6952
 - The color picker dialog will now show an alert if it is submitted with an invalid hex color code #TINY-2814
+- Fixed a bug where the `TableModified` event was not fired when adding a table row via the Tab key #TINY-7006
 
-## 5.7.1 - TBD
+## 5.7.1 - 2021-03-17
 
 ### Fixed
+- Fixed the `help` dialog incorrectly linking to the changelog of TinyMCE 4 instead of TinyMCE 5 #TINY-7031
 - Fixed a bug where error messages were displayed incorrectly in the image dialog #TINY-7099
 - Fixed an issue where URLs were not correctly filtered in some cases #TINY-7025
 - Fixed a bug where context menu items with names that contained uppercase characters were not displayed #TINY-7072
 - Fixed context menu items lacking support for the `disabled` and `shortcut` properties #TINY-7073
-- Fixed a regression whereby the width and height weren't set correctly when embedding content using the `media` dialog #TINY-7074
+- Fixed a regression where the width and height were incorrectly set when embedding content using the `media` dialog #TINY-7074
 
 ## 5.7.0 - 2021-02-10